Crate one_shot_mutex

One-shot locks that panic instead of (dead)locking on contention.

These locks allow no contention and panic instead of blocking on lock if they are already locked. This is useful in situations where contention would be a bug, such as in single-threaded programs that would deadlock on contention.

See the RawOneShotMutex and RawOneShotRwLock types for more information.

Structs

RawOneShotMutex

A one-shot mutex that panics instead of (dead)locking on contention.

RawOneShotRwLock

A one-shot readers-writer lock that panics instead of (dead)locking on contention.

Type Aliases

OneShotMutex

A lock_api::Mutex based on RawOneShotMutex.

OneShotMutexGuard

A lock_api::MutexGuard based on RawOneShotMutex.

OneShotRwLock

A lock_api::RwLock based on RawOneShotRwLock.

OneShotRwLockReadGuard

A lock_api::RwLockReadGuard based on RawOneShotRwLock.

OneShotRwLockUpgradableReadGuard

A lock_api::RwLockUpgradableReadGuard based on RawOneShotRwLock.

OneShotRwLockWriteGuard

A lock_api::RwLockWriteGuard based on RawOneShotRwLock.