one_shot_mutex

Struct RawOneShotRwLock

pub struct RawOneShotRwLock { /* private fields */ }

A one-shot readers-writer lock that panics instead of (dead)locking on contention.

This lock allows no contention and panics on lock_shared, lock_exclusive, lock_upgradable, and upgrade if it is already locked conflictingly. This is useful in situations where contention would be a bug, such as in single-threaded programs that would deadlock on contention.

Examples

use one_shot_mutex::OneShotRwLock;

static X: OneShotRwLock<i32> = OneShotRwLock::new(42);

// This is equivalent to `X.try_write().unwrap()`.
let x = X.write();

// This panics instead of deadlocking.
// let x2 = X.write();

// Once we unlock the mutex, we can lock it again.
drop(x);
let x = X.write();